#4614: gitmodules: Use https URLs for submodules
This means that if you clone with https://github.com/tenstorrent-metal/tt-metal,
the submodules will also use https. Slight removal of friction for
those not using SSH for github (cloning).

The alternative would be to use relative submodule paths, but that means
that anyone forking the tt-metal repo also needs to fork the submodules
for them to be possible to clone, which is awkward.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>