#5323: Convolutions of small size fail during parallelization calculations
Height Sharded fails because the code assumes the parallelization to be
at least 12 cores and then special-cases the overflow if it's not a
multiple of 12. But if the total number of par cores is < 12 to begin
with the code still assumes at least 12 cores and we end up with
rounding errors. The fix is to special case par for < 12 cores.