python_env: install pip dependencies with --no-compile
On a from-scratch `make build` + `make tests` it seems to save about 30s
(out of a ~12-13m build time) for me.

In general, this makes `pip install` not precompile the installed
modules. It makes for slightly slower invocation the first time they're
used, but many of our dependencies aren't used in the main build flow so
this seems to be a fair trade-off for initial build times.

Signed-off-by: Olof Johansson <olofj@tenstorrent.com>