#6048: add linear topology support to allgather for linear topology
Add all-gather topology info. The new line all-gather API is similar
to the ring all gather API. The arguments are the same with only the
op name being different: `line_all_gather`

Added concept (for linear) of duplicate full sends. For a linear
allgather, there is no wrap-around, so the full input tensor must
be sent in both directions. To simplify implementation/bringup time, the
existing ring all-gather was augmented such that the bulk of the
implementation was kept the same, but called in a loop - once for each
line direction (for ring we only run one loop but enable bidirectional).
For ring, two two loop iterations are for the "clockwise" and
"counter-clockwise" directions respectively. In loop iteration 0, all
workers operate in clockwise direction. For loop iteration 1, all
workers operate in counter-clockwise mode. Each loop iteration
instantiates a distinct set of workers.

Some refactoring was necessary to enable this:
- Added EDM arg generator component which simplifies edm
  arg generation to keep arg generation consistent from the
  caller in all_gather_op_multi_core. The caller API to connect
  workers to the EDM is simplified and EDM kernel generation can
  be done as a separate step.
  - This simplified setup - linear has non-uniform number of
    senders/receivers depending on which direction and chip we are
    currently at.
- Lots of changes to all_gather_op_multi_core to generalize further
  between linear and ring. Further refactoring coming down the pipe