#0: optimize allgather for small tensor sizes
For smaller tensor sizes, there is only a single packet sent through the
erisc data mover channel and that packet may be smaller in size than the
channel buffer. For those cases, the datamover channel buffer is shrunk
to be the same size as the packet. This can save a large amount of time.
For LLama, there are smaller allgathers of size 32x1024 which are
allgathered on dim=3. In those cases we can get up to a 2x improvement
for bfp8 and slightly less for fp16.