Generalizing sharding support in CCLs (#11074)
#10786: add generalized (page granularity) sharded tensor indexing to CCLs

Added a page index to noc address component, analogous to InterleavedAddrGen. In addition to this, templatized builders have been added to allow a kernel to invoke a generic AddrGen that will be selected based on a passed in `TensorMemoryLayout` argument. This way the op kernel code isn't required to change but can automatically support performance improvements, new sharding configurations, and new memory layouts.

As a results we now support:
- Width Sharded All Gather (Line and Ring)
- Height Sharded All Gather (Line and Ring)
- Block Sharded All Gather (Line and Ring)
- Width Sharded Reduce Scatter
- Block Sharded Reduce Scatter
- FullTensor bidirectional mode for sharded all-gather
  - Results in a 25% cycle count reduction for small all-gathers

Additionally, over 1000 lines of code have been able to be deleted.