Enable multi-buffer per channel in EDM (#11387)
#6300: Add multi-buffering per EDM channel

Adds the option to add a multiple buffers (e.g. double buffered) per EDM
channel. This is useful for improving performance of CCL operation.

To simplify the worker <-> EDM interface to allow a kernel to
automatically support multi-buffered channels, new adapter components
are added:
 - WorkerToEdmReader: for a worker pulling data from EDM
 - WorkerToEdmSender: for a worker pushing data to the EDM

These hide details such as buffer offsets in the channel and any other
details that may only be relevant to the EDM. Additionally, their use
encapsulates the worker <-> EDM data movement protocol, allowing future
low level changes to buffer layouts and allocations on the EDM without
requiring worker kernel changes.

As an a coinciding required step to enable this functionality the EDM
channel count limit has been lifted to unlimited (limited only to as
many buffers can fit into L1). This provides additional flexibility for
op writers and let's the `erisc_info::channels` to be shrunk back to
single entry.

Note that this commit only adds this feature, but does not yet enable it
for CCL ops.