#12387: Workaround to_layout for height sharded tensor
Conv2d will potentially trigger ttnn::to_layout for the input tensor in order to tilize the input. In cases where the input needs to be padded up to a multiple of 32 in order to be tilized, ttnn::tilize_with_val_padding will be triggered.

The problem is that ttnn::tilize_with_val_padding only supports width sharding, while height sharding is the most common sharding for Conv2d input.

To work around this issue, ttnn::to_layout is adjusted to first pad the input if it is height-sharded. After padding, ttnn::tilize is called, which does support height-sharded tensors.